Move theme to its own module and add spacing and padding
